Serial type fix

+++ tiny_dnn/activations/activation_layer.h
@@ -70,7 +70,7 @@ class activation_layer : public layer {
                            std::vector<tensor_t *> &out_data) override {
     const tensor_t &x = *in_data[0];
     tensor_t &y       = *out_data[0];
-    for_i(x.size(), [&](int i) { forward_activation(x[i], y[i]); });
+    for_i(x.size(), [&](size_t i) { forward_activation(x[i], y[i]); });
   }
 
   void back_propagation(const std::vector<tensor_t *> &in_data,
